chore: move to `lib`
BREAKING CHANGES:

* Implementation moved to `lib` directory
