+++ lib/contexts/diff.js
