+++ lib/diffpatcher.js
