+++ lib/filters/arrays.js
