+++ lib/filters/lcs.js
