+++ lib/filters/texts.js
