+++ lib/filters/trivial.js
