+++ lib/formatters/annotated.js
