+++ lib/formatters/console.js
