+++ lib/formatters/index.js
