+++ lib/pipe.js
