+++ lib/processor.js
